add access

+++ app/Http/Controllers/VoteController.php
@@ -6,15 +6,14 @@ use App\Http\Requests\VotesRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Auth\Access\AuthorizationException;
 use DCN\RBAC\Traits\HasRoleAndPermission;
+use DCN\RBAC\Exceptions\PermissionDeniedException;
 use DCN\RBAC\Contracts\HasRoleAndPermission as HasRoleAndPermissionContract;
 
 
 class VoteController extends ApiController implements HasRoleAndPermissionContract
 {
     use HasRoleAndPermission;
-
 
     /**
      * @param $votes array
@@ -38,7 +37,6 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
         }
         return $data;
     }
-
     /**
      * Display a listing of the resource.
      *
@@ -46,6 +44,11 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
      */
     public function index()
     {
+        $vote = new Vote();
+        $userAuth = Auth::user();
+        if (!($userAuth->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('index');
+
         $votes = Vote::all();
         $data = $this->getMetaData($votes);
         return $this->setStatusCode(200)->respond($data);
@@ -58,6 +61,11 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
      */
     public function store(VotesRequest $request)
     {
+        $vote = new Vote();
+        $userAuth = Auth::user();
+        if (!($userAuth->allowed('create.votes', $vote)))
+            throw new PermissionDeniedException('create');
+
         $vote = Vote::create($request->all());
         return $this->setStatusCode(201)->respond($vote);
     }
@@ -70,6 +78,10 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     public function show($id)
     {
         $vote = Vote::findOrFail($id);
+        $userAuth = Auth::user();
+
+        if (!($userAuth->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('view');
 
         $user = $vote->user()->first();
         $likeCount = $vote->likes()->count();
@@ -92,6 +104,11 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     public function update(VotesRequest $request, $id)
     {
         $vote = Vote::findOrFail($id);
+        $userAuth = Auth::user();
+
+        if (!($userAuth->allowed('update.votes', $vote)))
+            throw new PermissionDeniedException('update');
+
         $vote->update($request->all());
         return $this->setStatusCode(200)->respond($vote);
     }
@@ -104,6 +121,11 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     public function destroy($id)
     {
         $vote = Vote::findOrFail($id);
+        $userAuth = Auth::user();
+
+        if (!($userAuth->allowed('delete.votes', $vote)))
+            throw new PermissionDeniedException('delete');
+
         $vote->delete();
 
         if ($vote->trashed()) {
@@ -119,6 +141,11 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
         $user = User::findOrFail($userId);
         $votes = $user->votes()->get();
 
+        $vote = $votes[0];
+        $userAuth = Auth::user();
+        if (!($userAuth->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('index');
+
         if(!$votes){
             return $this->setStatusCode(200)->respond();
         }
@@ -129,6 +156,10 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     {
         $user = User::findOrFail($userId);
         $vote = $user->getVote($voteId);
+
+        $userAuth = Auth::user();
+        if (!($userAuth->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('view');
 
         if(!$vote){
             throw (new ModelNotFoundException)->setModel(Vote::class);
